--- a/src/layers/generic.cpp
+++ b/src/layers/generic.cpp
@@ -1,5 +1,5 @@
  
-/* All or part of this file was contributed by NVIDIA under license:
+/* Part of this file was contributed by NVIDIA under license:
  *   Copyright (C) 2020 NVIDIA Corporation
  *   SPDX-License-Identifier: MIT
  */