### Description

This PR splits out some small optimizations from PR #743.

Performance improvements from this PR as measured on a Titan V using a proxy transformer model:


Times with 1 Stream
Batch | Initial Time (s) | Current Time(s) | % Runtime reduction | Speedup factor
-- | -- | -- | -- | --
1 | 204.41 | 166.74 | 0.184286483 | 1.225920595
2 | 136.77 | 114.46 | 0.163120567 | 1.194915254
4 | 84.76 | 69.39 | 0.181335536 | 1.221501657
8 | 50.32 | 41.52 | 0.174880763 | 1.21194605
16 | 29.94 | 25.07 | 0.162658651 | 1.194256083
32 | 18.51 | 15.98 | 0.136682874 | 1.158322904
64 | 12.09 | 10.57 | 0.125723739 | 1.143803217
128 | 8.19 | 7.15 | 0.126984127 | 1.145454545
256 | 6.02 | 5.39 | 0.104651163 | 1.116883117


Times with two streams
Batch | Initial Time (s) | Current Time(s) | % Runtime reduction | Speedup factor
-- | -- | -- | -- | --
1 | 156.74 | 119.03 | 0.240589511 | 1.316810888
2 | 106.13 | 79.63 | 0.249693772 | 1.33278915
4 | 65.1 | 48.64 | 0.252841782 | 1.338404605
8 | 38.46 | 29.07 | 0.244149766 | 1.323013416
16 | 22.68 | 17.29 | 0.237654321 | 1.311740891
32 | 14.31 | 10.98 | 0.232704403 | 1.303278689
64 | 9.51 | 7.19 | 0.243953733 | 1.322670376
128 | 6.66 | 5.18 | 0.222222222 | 1.285714286
256 | 5.05 | 4.06 | 0.196039604 | 1.243842365




List of changes:
- Uses the per thread default stream for cublas
- Uses the strided batched gemm cublas call when possible for the batchedGemm. 
- In the general batchedGemm case, reduces the number of memcpy calls from 3 to 1.
- Rounds the width of input batches to a multiple of 8 when the GPU backend is being used. This is to enable better use of tensorcores on Volta architectures and newer
- Adds notices to the files changed

### How to test
I ran the regression tests. On Volta, some regression tests fail due to the additional use of tensor cores. This is because prior to CUDA 11, cublas does not use tensorcores if matrices fail to meet alignment requirements. This restriction is lifted in CUDA >= 11.

The differences in the float output from the failing regression tests are small. A list of them are provided here:
  - tests/interface/input-tsv/test_tsv_train_with_align.sh
  - tests/interface/input-tsv/test_tsv_train_with_align_and_weights.sh
  - tests/interface/input-tsv/test_tsv_train_with_align_and_weights_inputtypes.sh
  - tests/interface/input-tsv/test_tsv_train_with_align_pos0.sh
  - tests/interface/input-tsv/test_tsv_train_with_align_shuffle_in_ram.sh
  - tests/interface/input-tsv/test_tsv_train_with_align_stdin.sh
  - tests/scorer/nbest/test_compare_parallel_and_nbest.sh
  - tests/training/features/mixed-ensembles/test_ensemble_of_different_s2s.sh
  - tests/training/features/mixed-ensembles/test_ensemble_of_s2s_and_transformer.sh
  - tests/training/features/guided-alignment/test_guided_alignment_rnn.sh
  - tests/training/features/guided-alignment/test_guided_alignment_transformer.sh

OS: Ubuntu 18.04.3 LTS
Compiler gcc 7.5.0
nvcc version: 10.1.243

cmake command:

cmake .. -DCOMPILE_CPU=on -DCOMPILE_CUDA=on -DUSE_SENTENCEPIECE=on -DUSE_STATIC_LIBS=off -DCOMPILE_SERVER=off -DUSE_FBGEMM=on   -DCOMPILE_CUDA_SM35=off -DCOMPILE_CUDA_SM50=off -DCOMPILE_CUDA_SM60=off -DCOMPILE_CUDA_SM70=on -DCOMPILE_CUDA_SM75=off -DCOMPILE_TESTS=on


### Checklist

- [x] I have tested the code manually
- [x] I have run regression tests
- [x] I have read and followed CONTRIBUTING.md
- [ ] I have updated CHANGELOG.md